requirements.txt
Keras==2.1.5
tensorflow-gpu==1.4.0
torch==0.4.0
torchvision==0.2.0
Pillow==5.1.0
tensorboardX==1.2
